処理を行うメソッドとメッセージ付きのものに分けた
